letAPI/requirements.txt
click==7.1.2
DateTime==4.3
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector==2.2.9
PyJWT==1.7.1
pytz==2020.1
SQLAlchemy==1.3.18
uuid==1.30
Werkzeug==1.0.1
zope.interface==5.1.0
gunicorn
requests==2.23.0
oauth2==1.9.0.post1
